refactor: ScheduleCard 내부 기능 분리로 코드 품질 개선

+++ app/src/main/java/com/example/myschedule/components/ScheduleCard.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
@@ -12,11 +11,11 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.example.myschedule.R
 import com.example.myschedule.data.WeatherCode
 import com.example.myschedule.data.database.entity.Schedule
-import com.example.myschedule.schedulecreate.titletimeinput.ScheduleTime
 import com.example.myschedule.ui.theme.*
 
 @Composable
@@ -33,12 +32,12 @@ fun ScheduleCard(
     Card(
         colors = CardDefaults.cardColors(White),
         border = BorderStroke(1.dp, LightGray),
-        modifier = Modifier.fillMaxWidth()
+        modifier = Modifier.fillMaxWidth().wrapContentHeight()
     ) {
-        Row(modifier = Modifier.height(IntrinsicSize.Min)) {
+        Row(verticalAlignment = Alignment.CenterVertically) {
             Box(
                 contentAlignment = Alignment.Center,
-                modifier = Modifier.width(48.dp).fillMaxHeight()
+                modifier = Modifier.size(48.dp)
             ) {
                 Icon(
                     painter = painterResource(R.drawable.star),
@@ -48,8 +47,6 @@ fun ScheduleCard(
                 )
             }
 
-            Spacer(modifier = Modifier.width(8.dp))
-
             Column {
                 Text(
                     text = schedule.title,
@@ -58,98 +55,72 @@ fun ScheduleCard(
                 )
 
                 Text(
-                    text = scheduleTimeRange(schedule.startTime, schedule.endTime),
+                    text = stringResource(
+                        id = R.string.schedule_time_range,
+                        schedule.startTime.amPm, schedule.startTime.hour, schedule.startTime.minute,
+                        schedule.endTime.amPm, schedule.endTime.hour, schedule.endTime.minute
+                    ),
                     color = Gray,
                     style = MaterialTheme.typography.bodyMedium
                 )
             }
 
-            if (showWeather || showOptions) {
-                Spacer(modifier = Modifier.weight(1f))
+            Spacer(modifier = Modifier.weight(1f))
 
-                WeatherOrOption(showWeather, weatherCode, showOptions, onEditClick, onDeleteClick)
+            if (showWeather) {
+                WeatherImage(weatherCode)
+            } else if (showOptions) {
+                OptionIcon(onEditClick, onDeleteClick)
             }
         }
     }
 }
 
-fun scheduleTimeRange(startTime: ScheduleTime, endTime: ScheduleTime): String {
-    return "${startTime.amPm} ${startTime.hour}시 ${startTime.minute.toString().padStart(2, '0')}분" +
-            " ~ " +
-            "${endTime.amPm} ${endTime.hour}시 ${endTime.minute.toString().padStart(2, '0')}분"
+@Composable
+fun WeatherImage(weatherCode: WeatherCode?) {
+
+    Box(
+        contentAlignment = Alignment.CenterEnd,
+        modifier = Modifier.fillMaxSize().padding(end = 8.dp)
+    ) {
+        val imgResId = when (weatherCode) {
+            WeatherCode.SUNNY -> R.drawable.sunny
+            WeatherCode.SUNNY_WITH_CLOUD -> R.drawable.little_cloudy
+            WeatherCode.CLOUDY -> R.drawable.cloudy
+            WeatherCode.RAINY -> R.drawable.rainy
+            WeatherCode.SNOWY -> R.drawable.snowy
+            else -> R.drawable.close
+        }
+
+        Image(
+            painter = painterResource(imgResId),
+            contentDescription = null,
+            modifier = Modifier.size(32.dp)
+        )
+    }
 }
 
 @Composable
-fun WeatherOrOption(
-    showWeather: Boolean,
-    weatherCode: WeatherCode?,
-    showOptions: Boolean,
-    onEditClick: (() -> Unit)? = null,
-    onDeleteClick: (() -> Unit)? = null
-) {
+fun OptionIcon(onEditClick: (() -> Unit)? = null, onDeleteClick: (() -> Unit)? = null) {
 
     var dropDownExpanded by remember { mutableStateOf(false) }
 
-    if (showWeather) {
-        Box(
-            contentAlignment = Alignment.CenterEnd,
-            modifier = Modifier.fillMaxSize().padding(end = 8.dp)
-        ) {
-            when (weatherCode) {
-                WeatherCode.SUNNY -> Image(
-                    painter = painterResource(R.drawable.sunny),
-                    contentDescription = null,
-                    modifier = Modifier.size(32.dp)
-                )
-
-                WeatherCode.SUNNY_WITH_CLOUD -> Image(
-                    painter = painterResource(R.drawable.little_cloudy),
-                    contentDescription = null,
-                    modifier = Modifier.size(32.dp)
-                )
-
-                WeatherCode.CLOUDY -> Image(
-                    painter = painterResource(R.drawable.cloudy),
-                    contentDescription = null,
-                    modifier = Modifier.size(32.dp)
-                )
-
-                WeatherCode.RAINY -> Image(
-                    painter = painterResource(R.drawable.rainy),
-                    contentDescription = null,
-                    modifier = Modifier.size(32.dp)
-                )
-
-                WeatherCode.SNOWY -> Image(
-                    painter = painterResource(R.drawable.snowy),
-                    contentDescription = null,
-                    modifier = Modifier.size(32.dp)
-                )
-
-                else -> Icon(
-                    imageVector = Icons.Default.Close,
-                    contentDescription = null,
-                    tint = Black,
-                    modifier = Modifier.size(28.dp)
-                )
-            }
+    Box(contentAlignment = Alignment.CenterEnd) {
+        IconButton(onClick = { dropDownExpanded = true }) {
+            Icon(
+                imageVector = Icons.Default.MoreVert,
+                contentDescription = null,
+                tint = Black
+            )
         }
-    } else if (showOptions && onEditClick != null && onDeleteClick != null) {
-        Box(contentAlignment = Alignment.CenterEnd) {
-            IconButton(onClick = { dropDownExpanded = true }) {
-                Icon(
-                    imageVector = Icons.Default.MoreVert,
-                    contentDescription = null,
-                    tint = Black
-                )
-            }
 
-            DropdownMenu(
-                expanded = dropDownExpanded,
-                onDismissRequest = { dropDownExpanded = false },
-                shape = RoundedAllCornerShape,
-                containerColor = White,
-            ) {
+        DropdownMenu(
+            expanded = dropDownExpanded,
+            onDismissRequest = { dropDownExpanded = false },
+            shape = RoundedAllCornerShape,
+            containerColor = White,
+        ) {
+            if (onEditClick != null) {
                 DropdownMenuItem(
                     onClick = {
                         onEditClick()
@@ -157,7 +128,9 @@ fun WeatherOrOption(
                     },
                     text = { Text("수정") }
                 )
+            }
 
+            if (onDeleteClick != null) {
                 DropdownMenuItem(
                     onClick = {
                         onDeleteClick()
